src/test/java/com/lj/app/core/common/security/DESCoderTest.java
package com.lj.app.core.common.security;

import static org.junit.Assert.assertEquals;

import org.apache.commons.codec.binary.Base64;
import org.junit.Test;

public class DESCoderTest {

	@Test
	public void test() throws Exception{
		String inputStr = "DES";
		byte[] inputData = inputStr.getBytes();
		System.out.println("原文:"+inputStr);
		
		byte key[] = DESCoder.initKey();
		System.out.println("密钥:" + Base64.encodeBase64String(key));
		
		inputData = DESCoder.encrypt(inputData, key);
		System.out.println("加密后密文:" + Base64.encodeBase64String(inputData));
		
		byte []outputData = DESCoder.encrypt(inputData, key);
		String outputStr = new String(outputData);
		System.out.println("解密后原文:" + outputStr);
		
		assertEquals(inputStr,outputStr);
	}

}
